Remove form-specific and alias logic
