--- app/services/ocr/textract/aliases/consecutive_aliases.yaml
@@ -1,47 +1,0 @@
-"nombre(s) (sin abreviaturas)":
-  - nombre
-"apellido paterno":
-  - a_paterno
-"apellido materno":
-  - a_materno
-"telefono de casa":
-  - tel_casa
-"telefono celular":
-  - celular
-"fecha de nacimiento":
-  - fecha_nacimiento
-  - fecha de nacimiento
-"pais de nacimiento":
-  - pais_nacimiento
-  - pais de nacimiento
-"nacionalidad":
-  - nacionalidad
-"genero":
-  - sexo
-"e-mail":
-  - correo electronico
-  - email
-"tipo de identificacion":
-  - tipo_identificacion
-"folio":
-  - folio
-"estado civil":
-  - estado_civil
-"regimen matrimonial":
-  - regimen_matrimonial
-"domicilio (calle y numero exterior e interior)":
-  - domicilio_calle
-"colonia/urbanizacion":
-  - colonia_urbanizacion
-"delegacion/municipio/demarcacion politica":
-  - delegacion_municipio_demarcacion_politica
-"ciudad/poblacion":
-  - ciudad_poblacion
-"entidad federativa/estado":
-  - entidad_federativa_estado
-"pais":
-  - pais
-"cp":
-  - codigo_postal
-"anos de residencia":
-  - anos_residencia
