+++ app/services/postprocessors/generic_postprocessor.py
@@ -6,4 +6,8 @@ class GenericPostProcessor:
 
     def process(self, raw_fields: dict) -> dict:
         """Aplica limpieza genérica a los campos extraídos."""
-        return self.cleaner.transform(raw_fields)
+        cleaned = self.cleaner.transform(raw_fields)
+        checklist = [k for k, v in cleaned.items() if isinstance(v, str) and v.lower() == "sí"]
+        if checklist:
+            cleaned["checklist"] = checklist
+        return cleaned
